implement only APIs to service file
